backend/requirements.txt
# DOER Platform - Python Dependencies
# 
# Installation:
#   python -m venv venv
#   source venv/bin/activate  (or venv\Scripts\activate on Windows)
#   pip install -r requirements.txt
#
# PRODUCTION NOTES:
# - Pin exact versions for reproducibility
# - Add asyncpg for PostgreSQL support
# - Add boto3 for S3 integration
# - Add celery for background tasks
# - Add redis for caching

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.25
aiosqlite==0.19.0
greenlet==3.3.1

# PRODUCTION: Uncomment for PostgreSQL
# asyncpg==0.29.0
# alembic==1.13.1  # Database migrations

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# File Handling
# ============================================================================
python-multipart==0.0.6
PyPDF2==3.0.1
Pillow==10.2.0

# PRODUCTION: Uncomment for OCR
# pytesseract==0.3.10

# ============================================================================
# HTTP Client (for AI APIs)
# ============================================================================
httpx==0.26.0

# ============================================================================
# PRODUCTION DEPENDENCIES (uncomment as needed)
# ============================================================================

# AWS S3 Integration
# boto3==1.34.0

# Redis Cache & Rate Limiting
# redis==5.0.1
# fastapi-limiter==0.1.6

# Background Tasks
# celery==5.3.6

# Error Tracking
# sentry-sdk[fastapi]==1.39.1

# Metrics & Monitoring
# prometheus-fastapi-instrumentator==6.1.0

# SMS/WhatsApp (Twilio)
# twilio==8.11.0

# ============================================================================
# Development & Testing
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing FastAPI

# Code Quality
black==24.1.0
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
